Add reference update and delete
